requirements.txt
peewee~=3.16.3
python-telegram-bot ~=20.4
psycopg2~=2.9.7